requirements.txt
selenium
webdriver-manager
python-dotenv
undetected-chromedriver
openpyxl